requirements.txt
streamlit
openai
langchain
langchain-community
tiktoken
PyPDF2
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
faiss-cpu